+++ src/obr/OpenFOAM/case.py
@@ -208,8 +208,10 @@ class OpenFOAMCase(BlockMesh):
             self.file_dict[rel_path] = file
         for file, rel_path in self.config_files_in_folder(self.system_include_folder):
             self.file_dict[rel_path] = file
-        for file, rel_path in self.config_files_in_folder(self.const_polyMesh_folder):
-            self.file_dict[rel_path] = file
+        # TODO dont try to create File object for polyMesh files because that might
+        # take very long
+        # for file, rel_path in self.config_files_in_folder(self.const_polyMesh_folder):
+        #     self.file_dict[rel_path] = file
         return list(self.file_dict.keys())
 
     def get(self, key: str) -> Union[File, None]:
@@ -227,7 +229,21 @@ class OpenFOAMCase(BlockMesh):
         return logged_execute(operation, self.path, self.job.doc)
 
     def decomposePar(self, args={}):
-        """Sets decomposeParDict and calls decomposePar"""
+        """Sets decomposeParDict and calls decomposePar. If no decomposeParDict exists a new one
+        gets created"""
+
+        if not self.decomposeParDict:
+            decomposeParDictFile = Path(self.system_folder / "decomposeParDict")
+            with open(decomposeParDictFile, "a") as fh:
+                # call get to trigger read
+                self.controlDict.update()
+                fh.write("".join(self.controlDict.of_comment_header))
+                fh.write("".join(self.controlDict.of_header))
+                fh.write("\n")
+            self.decomposeParDict = File(
+                folder=self.system_folder, file="decomposeParDict", job=self.job
+            )
+
         method = args["method"]
         if method == "simple":
             if not args.get("numberOfSubDomains"):
@@ -243,7 +259,7 @@ class OpenFOAMCase(BlockMesh):
                 {
                     "method": method,
                     "numberOfSubdomains": numberSubDomains,
-                    "coeffs": {"n": coeffs},
+                    "simpleCoeffs": {"n": coeffs},
                 }
             )
         else: