+++ src/obr/core/core.py
@@ -6,12 +6,12 @@ import hashlib
 import logging
 import json
 
-
 from pathlib import Path
 from subprocess import check_output
-from typing import Union
+from typing import Union, Generator
 from datetime import datetime
 from signac.contrib.job import Job
+from copy import deepcopy
 
 # these are to be replaced with each other
 SIGNAC_PATH_TOKEN = "_dot_"
@@ -164,8 +164,8 @@ def get_mesh_stats(owner_path: str) -> dict:
                 if is_foamFile and "note" in line:
                     found_note = line
         note_line = found_note
-        nCells = int(re.findall("nCells:([0-9]+)", note_line)[0])
-        nFaces = int(re.findall("Faces:([0-9]+)", note_line)[0])
+        nCells = int(re.findall("nCells:[ ]*([0-9]+)", note_line)[0])
+        nFaces = int(re.findall("Faces:[ ]*([0-9]+)", note_line)[0])
     return {"nCells": nCells, "nFaces": nFaces}
 
 
@@ -195,11 +195,14 @@ def merge_job_documents(job: Job):
     job.doc = {"data": merged_data, "history": merged_history, "cache": cache}
 
 
-def get_latest_log(job: Job):
-    """find latest"""
+def get_latest_log(job: Job) -> str:
+    """Find latest log in job.id/case/folder
+
+    Returns: path to latest solver log
+    """
     from ..OpenFOAM.case import OpenFOAMCase
 
-    case_path = Path(job.path + "case")
+    case_path = Path(job.path + "/case")
     # in case obr status is called directly after initialization
     # this would also fail if there was no case directory
     if not case_path.exists():
@@ -210,9 +213,52 @@ def get_latest_log(job: Job):
 
     history = job.doc["history"]
     for entry in history[:-1]:
-        if entry.get("cmd", "").startswith(solver):
+        if solver in entry.get("cmd", ""):
             return entry["log"]
     return ""
+
+
+def get_timestamp_from_log(log) -> str:
+    """gets the timestamp part from an log file"""
+    log_name = Path(log).stem
+    before = log_name.split("_")[0]
+    return log_name.replace(before + "_", "")
+
+
+def find_solver_logs(job: Job) -> Generator[tuple, None, None]:
+    """Find and return all solver log files, campaign info and tags from job instances"""
+    case_path = Path(job.path)
+    if not case_path.exists():
+        return
+
+    root, campaigns, _ = next(os.walk(case_path))
+
+    def find_tags(path: Path, tags: list, tag_mapping):
+        """Recurses into subfolders of path until a system folder is found
+
+        Returns:
+          Dictionary mapping paths to tags -> tag
+        """
+        _, folder, _ = next(os.walk(path))
+        is_case = len(folder) == 0
+        if is_case:
+            tag_mapping[str(path)] = tags
+        else:
+            for f in folder:
+                tags_copy = deepcopy(tags)
+                tags_copy.append(f)
+                find_tags(path / f, tags_copy, tag_mapping)
+        return tag_mapping
+
+    for campaign in campaigns:
+        # check if case folder
+        tag_mapping = find_tags(case_path / campaign, [], {})
+
+        for path, tags in tag_mapping.items():
+            root, _, files = next(os.walk(path))
+            for file in files:
+                if "Foam" in file and file.endswith("log"):
+                    yield f"{root}/{file}", campaign, tags
 
 
 def execute(steps: list[str], job) -> bool:
@@ -259,11 +305,6 @@ def modifies_file(fns):
         unlink(fns)
 
 
-def check_log_for_success(log: Path) -> bool:
-    res = check_output(["tail", "-n", "2", log], text=True)
-    return ("Finalising" in res) or ("End" in res)
-
-
 def writes_files(fns):
     """check if this job modifies a file, thus it needs to unlink
     and copy the file if it is a symlink
@@ -302,5 +343,4 @@ def map_view_folder_to_job_id(view_folder: str) -> dict[str, str]:
                 # folder
                 if path.absolute().is_relative_to(base.absolute()):
                     ret[job_id] = str(path.absolute().relative_to(base.absolute()))
-                folder.pop(i)
     return ret